requirements.lock.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile --universal --strip-extras --output-file requirements.lock.txt requirements.txt
cffi==2.0.0 ; sys_platform == 'win32'
    # via clr-loader
click==8.3.3
    # via -r requirements.txt
clr-loader==0.2.10 ; sys_platform == 'win32'
    # via pythonnet
colorama==0.4.6 ; sys_platform == 'win32'
    # via click
comtypes==1.4.16 ; sys_platform == 'win32'
    # via
    #   -r requirements.txt
    #   pycaw
paho-mqtt==2.1.0
    # via -r requirements.txt
psutil==7.2.2
    # via
    #   -r requirements.txt
    #   pycaw
pycaw==20251023 ; sys_platform == 'win32'
    # via -r requirements.txt
pycparser==3.0 ; implementation_name != 'PyPy' and sys_platform == 'win32'
    # via cffi
pythonnet==3.0.5 ; sys_platform == 'win32'
    # via -r requirements.txt
sd-notify==0.1.0 ; sys_platform == 'linux'
    # via -r requirements.txt
